feat: gemini fallback + large file refactor
