+++ backend/pyproject.toml
@@ -15,6 +15,8 @@ dependencies = [
     "python-dotenv>=1.0.0",
     "pydantic>=2.10.0",
     "jinja2>=3.1.6",
+    "google-generativeai>=0.8.6",
+    "langchain-text-splitters>=1.1.0",
 ]
 
 [build-system]
